minor bug fixed(5/15/fri)

+++ CarDirection.ino
@@ -1,3 +1,3 @@
-void fx_CarDirection(int wheel_deg) {
+void fx_CarDirection(double wheel_deg) {
   servo_CarDirection.write(wheel_deg+90+CarDir_cali);                           // tell SG90 to turn a certain angle (e.g. turn right 60deg, then it sends a 150 signal) }                                                   // wait until the needed angle is reached
 }
